fix: resolve CI/CD build and test failures
- Remove duplicate repository files in places domain
- Fix trips domain model duplication
- Fix user repository search query parameters
- Add missing mock expectation for GetByUsername in tests
- Clean up backup files
